# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow API
  version: 0.1.0
  description: |
    REST API for the Stellar Trust Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.stellar-trust-escrow.example.com`
    - **Mainnet**: `https://api.stellar-trust-escrow.example.com`
  contact:
    name: Stellar Trust Escrow
    url: https://stellar-trust-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/Stellar-Trust-Escrow/stellar-trust-escrow
  x-docs-url: https://docs.stellar-trust-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.stellar-trust-escrow.example.com
    description: Testnet
  - url: https://api.stellar-trust-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/escrows:
    get:
      tags: [Escrows]
      summary: List escrows with filters and pagination
      operationId: listEscrows
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          description: Single value or comma-separated list
          schema:
            type: string
            enum: [Active, Completed, Disputed, Cancelled]
          example: Active,Completed
        - name: client
          in: query
          description: Filter by client Stellar address
          schema:
            type: string
            example: GDKHX...QZ4
        - name: freelancer
          in: query
          description: Filter by freelancer Stellar address
          schema:
            type: string
            example: GDKHX...QZ4
        - name: search
          in: query
          description: Substring search on escrow ID or address
          schema:
            type: string
        - name: minAmount
          in: query
          description: Minimum totalAmount (XLM string)
          schema:
            type: string
            example: '100'
        - name: maxAmount
          in: query
          description: Maximum totalAmount (XLM string)
          schema:
            type: string
            example: '5000'
        - name: dateFrom
          in: query
          description: 'ISO date — createdAt >= dateFrom'
          schema:
            type: string
            format: date
            example: '2024-01-01'
        - name: dateTo
          in: query
          description: 'ISO date — createdAt <= dateTo (end of day)'
          schema:
            type: string
            format: date
            example: '2024-12-31'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Paginated escrow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEscrows'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/disputes:
    get:
      tags: [Disputes]
      summary: List disputes with pagination
      operationId: listDisputes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/reputation/leaderboard:
    get:
      tags: [Reputation]
      summary: Top users by reputation score
      operationId: getLeaderboard
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeaderboard'
...
  /api/notifications/events:
    post:
      tags: [Notifications]
      summary: Enqueue an email notification event
      operationId: enqueueNotification
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType, data]
              properties:
                eventType:
                  type: string
                  enum:
                    - escrow.status_changed
                    - milestone.completed
                    - dispute.raised
                  example: milestone.completed
                data:
                  type: object
                  required: [recipients]
                  properties:
                    recipients:
                      type: array
                      items:
                        type: string
                        format: email
                      example: ['user@example.com']
                    escrowId:
                      type: string
                    milestoneId:
                      type: string
      responses:
        '202':
          description: Event accepted for delivery
          content:
            application/json:
              schema:
                type: object
                properties:
                  enqueued:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List active webhook subscriptions
      operationId: listWebhookSubscriptions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        url:
                          type: string
                        eventTypes:
                          type: array
                          items:
                            type: string
                        isActive:
                          type: boolean
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
    post:
      tags: [Webhooks]
      summary: Register a webhook subscription
      operationId: createWebhookSubscription
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                eventTypes:
                  type: array
                  items:
                    type: string
              required: [url, eventTypes]
      responses:
        '201':
          description: Subscription created with secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      url:
                        type: string
                      eventTypes:
                        type: array
                        items:
                          type: string
                      secret:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/events:
    get:
      tags: [Events]
      summary: List all indexed on-chain events
      operationId: listEvents
      security: []
      parameters:
        - name: eventType
          in: query
          schema:
            type: string
          example: escrow_created
        - name: escrowId
          in: query
          schema:
            type: string
        - name: fromLedger
          in: query
          schema:
            type: integer
          description: Inclusive lower bound ledger sequence
        - name: toLedger
          in: query
          schema:
            type: integer
          description: Inclusive upper bound ledger sequence
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
...
  /api/kyc/admin:
    get:
      tags: [KYC]
      summary: Admin — list all KYC records
      operationId: adminListKyc
      security:
        - AdminAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, Init, Processing, Approved, Declined]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated KYC records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedKycRecords'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users with pagination and search
      operationId: adminListUsers
      security:
        - AdminAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminUsers'
...
  /api/admin/disputes:
    get:
      tags: [Admin]
      summary: List all disputes with pagination
      operationId: adminListDisputes
      security:
        - AdminAuth: []
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/audit:
    get:
      tags: [Audit]
      summary: Search audit logs with filters
      operationId: searchAuditLogs
      security:
        - AdminAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          example: escrow
        - name: action
          in: query
          schema:
            type: string
          example: created
        - name: actor
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
        - name: from
          in: query
          description: ISO date/time lower bound
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: ISO date/time upper bound
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, totalAmount, status]
        default: createdAt
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    Escrow:
      type: object
      properties:
        id:
          type: string
          example: escrow_abc123
        client:
          type: string
          example: GDKHX...QZ4
        freelancer:
          type: string
          example: GDKHX...QZ4
        totalAmount:
          type: string
          example: '1000'
        status:
          type: string
          enum: [Active, Completed, Disputed, Cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginatedEscrows:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Escrow'
...
    UserProfile:
      type: object
      properties:
        address:
          type: string
        email:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [Pending, Init, Processing, Approved, Declined]
        reputation:
          type: object
          properties:
            score:
              type: number
            totalReviews:
              type: integer
            completionRate:
              type: number
        escrowStats:
          type: object
          properties:
            asClient:
              type: integer
            asFreelancer:
              type: integer
            completed:
              type: integer
            disputed:
              type: integer
        createdAt:
          type: string
          format: date-time
...
    Dispute:
      type: object
      properties:
        id:
          type: string
        escrowId:
          type: string
        raisedBy:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [Open, Resolved]
        resolution:
          type: object
          nullable: true
          properties:
            clientAmount:
              type: string
            freelancerAmount:
              type: string
            notes:
              type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PaginatedDisputes:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Dispute'
    ReputationRecord:
      type: object
      properties:
        address:
          type: string
        score:
          type: number
        totalReviews:
          type: integer
        positiveReviews:
          type: integer
        negativeReviews:
          type: integer
        completionRate:
          type: number
        avgResponseTimeHours:
          type: number
          nullable: true
        rank:
          type: integer
          nullable: true
    PaginatedLeaderboard:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ReputationRecord'
...
    IndexedEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ledger:
          type: integer
        txHash:
          type: string
        eventType:
          type: string
        escrowId:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
    PaginatedEvents:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/IndexedEvent'
    KycStatus:
      type: object
      properties:
        address:
          type: string
        status:
          type: string
          enum: [Pending, Init, Processing, Approved, Declined]
        reviewResult:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedKycRecords:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/KycStatus'
...
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            banStatus:
              type: string
              enum: [active, suspended, banned]
            banReason:
              type: string
              nullable: true
            bannedAt:
              type: string
              format: date-time
              nullable: true
    PaginatedAdminUsers:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AdminUserDetail'
...
    PaginatedAuditLogs:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AuditLogEntry'
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        action:
          type: string
        actor:
          type: string
        resourceId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'